The CMakeLists were not making Visual C++ happy. This has been fixed. Several types issues have been fixed. The /Wp64 option of Visual C++ has been turned off, since it reports wrongly some things as errors (dixit Microsoft and my own experience).

+++ meteoio/libsmet.cc
@@ -627,12 +627,12 @@ void SMETWriter::check_formatting()
 	}
 }
 
-void SMETWriter::set_width(const std::vector<size_t>& vec_width)
+void SMETWriter::set_width(const std::vector<int>& vec_width)
 {
 	ascii_width = vec_width;
 }
 
-void SMETWriter::set_precision(const std::vector<size_t>& vec_precision)
+void SMETWriter::set_precision(const std::vector<int>& vec_precision)
 {
 	ascii_precision = vec_precision;
 }
